rename aux_client.py to client.py
